Add script to calculate distances between amino acid sequences
This interface is a prototype for a future modular augur interface that will
replace part of the current augur "scores" calculations. This script
specifically enables the calculation of amino acid distances using predefined
masks. For flu builds, this means epitope, non-epitope, and "receptor binding
site" mutations can be counted for each node relative to the root with a single
command line script.