Reconstruct nucleotide and amino acid sequences for a given tree
Use the sequences exported for the root of a tree by the augur export command
and the corresponding tree with mutations annotated per node to reconstruct each
node's nucleotide and amino acid sequence.

The bulk of the work is performed by a standalone function that could eventually
be integrated into augur. The function can be executed from the command line
through the script's command line interface to produce a sequence-annotated tree
JSON.