Use raw modular augur outputs to reconstruct sequences
Instead of using output from augur export that is ready for auspice display, use
the information about root sequencs and mutations that is already available from
intermediate modular augur steps. This approach removes a circular dependency of
reconstructed sequences and augur export and also makes the reconstruction code
and data a lot simpler.

This new approach requires using the latest version of augur in the Snakefile.