Identify non-overlapping clades with continuity between timepoints
This approach identifies all clades that are alive across a range of timepoints
and selects only those clades whose parent is not selected in the current
timepoint. The benefit of this approach is that large clades that fix are
automatically split into their smaller subclades and small clades that die are
automatically omitted at the appropriate time.