Visualize model input data frame and clade frequencies
To assist with debugging model fit and exploration of predictor relationships,
this notebook includes plots of standardized predictor values and frequency
trajectories of clades selected for fitting. These latter plots help reveal the
difficulty of fitting a model by showing the start and end frequencies of clades
in a prediction interval.

The included frequencies are estimated with a narrow KDE bandwidth of 3
months. The included model inputs only consider the cross-immunity amplitude
predictor (ep_x).