Final changes to get fitness models running with modular augur
Moves remaining rules from original Snakefile into the corresponding quality
control and fitness model files and fixes some rules that were broken in the
modular augur transition. Notably, the LBI predictor used by the frequency
bandwidth analysis does not work as expected with the quickstart data and will
need to be fixed in the future.

Builds now use modular augur (installed as a Python module) and follow the
convention of expecting fauna to be installed in a sibling directory. This
latter convention will (hopefully) be replaced again with a proper fauna
module someday.

Remove submodules for fauna and augur and note that fauna should be installed as
a sibling directory.