Handle case when no train or test clades are available
Provides better debugging to the user about which timepoints are used for
training and testing and also skips any train/test set for which no clades exist
for training or testing with a warning to the user. This case should disappear
if clades are defined by haplotype in the future.