Ensure no future sequences are included in builds
Fixes a bug in the logic associated with filtering strains by date which was
causing strains from outside the requested time interval to be included in the
final output. This bug was due to an off-by-one error in parsing of
months. Although this commit fixes this specific error, it also moves the
filtering of strains by date to the beginning of the flu subsampling function
and uses numeric floating point dates. This approach allows users to specify
time intervals to the day and have the subsampling respect that precise of a
request.

Closes #19.