Collect predictor JSONs into per-timepoint data frames
Adds distance metrics like epitope, non-epitope, and RBS mutations to the
modular augur build and uses the new node data to table script to collect these
JSONs into a single table with a timepoint column and one row per strain. This
approach will allow multiple timepoints to be aggregated into a similar data
frame to be used as input to the fitness model.