Calculate train-validate timepoints
Adds a new function to the fitness model module to calculate valid
train-validate timepoints based on a fixed-size training window (measured in
years) and a delta time to project into the future (also in years). The function
returns a list of dictionaries indexed by "train" and "validate" keys. This
output will be used by Snakemake to pass arguments to the command line utility
for standardizing predictors and also to define the expected model output JSONs
when we start training one model per combinatio of predictors and train-validate interval.