Add support for multiple lineages
Build trees for all defined segments for all defined lineages with support for
lineage-specific outputs. This commit also adds reference FASTAs, lists of
reference viruses to include in builds, outliers, and site masks that are
currently defined for all lineages in the standard seasonal flu builds.

This commit also updates the titer download rule to support passage and assay
specification in addition to lineage specification.

Closes #30.