Implement fitness model interface
Includes a command line interface for fitting models with cross-validation, a
generic function for cross-validation, and a generic exponential growth model
that is patterned after the scikit-learn model interfaces.

Models are fit by minimization of sum of squared errors across timepoints and
clades using data frames of independent and dependent variables as
inputs. Models are stored in JSON format with model coefficients, errors per
training window, and raw data used for training and validation.

One major interface change here is the use of delta months (integers) in most
user-facing scripts instead of delta time (years as floats). This allows date
offsets to play nicely in pandas and also allows for delta times values that are
less than one year without resorting to fractions and the associated rounding
frustrations associated with that approach.

Closes #18.
Closes #23.