Add cost function based on information gain
Adds a cost function for negative information gain that can be optimized through
scipy's minimize function in the same way sum of squared errors is
minimized. The final scores reported for information gain are divided by the
number of samples, so the reported value for cross-validation error is the mean
negative information gain per clade. Optimal models then have the smallest value
and preferably negative. Information gain requires the addition of pseudocounts
to frequencies, so support for adding pseudocounts is included in this commit,
too.

This commit also refactors the cost function logic to enable arbitrary
functions.