Add support for SHA hash clade ids
Although the default strings of mutations are the most biologically
interpretable clade ids, they are impossible to visually compare or debug. This
commit implements Trevor's original idea of hashing the mutation strings and
reporting only the first n characters of that hash as a unique id. This has made
debugging much easier.