Identify clades by mutations relative to a shared reference
Previously clades had been identified by mutations between ancestral or observed
viruses and the MRCA of the tree from each timepoint. These mutations were used
to link clade assignments between timepoints and enable tracking of clade
frequencies over time. However, the addition of new viruses to each timepoint
led to variability in the MRCA ancestral sequence reconstruction that broke
these clade linkages.

This commit resolves this issue by calculating clade haplotypes relative to a
shared reference sequence that does not depend on any timepoint-specific
data. The reference sequence used is the same used to produce the initial
multiple sequence alignment that the tree is based on. This commit also reduces
the number of tips required per clade from 10 to 2. This number was empirically
determined to be the minimum number required to ensure that the same clades
would be present in consecutive timepoints despite any variability in ancestral
sequence reconstruction.

Closes #40.