Create delta frequency predictor
Adds script, rules, and config to calculate the change in frequency for clades
between the current timepoint and a predefined number of pivot steps back in the
past. The "delta_frequency" is technically the change in frequency over time or
the slope of the observed frequency trajectory per clade. These per-clade values
are passed back down the tree to their corresponding tips for use as a fitness
metric in forecasting models and display in auspice.

Closes #34.