Updated and expanded summary figures for single-tree models
For comparison of single-tree model results to current multi-tree models, the
original model summary notebook now includes plots for MAE by model and by model
and timepoint against the naive. Although I had previously forgotten this, the
ep_x model had not performed better than the naive model in the past with single
high-density trees either, so it is possible that other model parameters (or
even bugs) created the memory/perception of ep_x's better performance. This
needs to be investigated more.