Calculate correct estimated distance to future
Updates model fitting script and rules to pass raw pairwise distances to model
fitting class so estimated distances to the future per virus can be calculated
correctly. The difference here is that the previous calculations used the
weighted distance to the present times the projected frequency of each virus
when the order of these operations should have been the estimated weighted
distance to the future (based on projected frequencies of viruses at the current
timepoint) times the current frequency of each virus.

Since we do not fit models with this estimated value, it is only calculated
after the model is already fit such that the data frames in the JSON output
contain the observed and estimated weighted distances to the future for
downstream comparisons.