Added XML config for santa-sim for influenza-like simulations
This config simulates HA evolution starting with the Beijing 1992 reference
sequence and applying purifying selection to SigPep, HA1, and HA2 and positive
and exposure-dependent selection to HA1 epitope sites based on the Wolf
sites. Since santa-sim is amino-acid-aware and supports seeding by an originally
functional HA sequence, we can use the exact set of epitope and non-epitope
sites here that we use in our analysis of natural sequences.