Update modular augur rules for simulations with SANTA-SIM sequences
New simulated sequences can be translated to amino acid sequences and passed
through most of the same pipeline as the historical HA sequences. This allows
for inspection of the exact same non-epitope and epitope sites in both simulated
and historical sequences and comparison of model results where we know what the
evolutionary pressures are in the simulated sequences.