Increase sampling frequency, decrease total samples, export fitness
Uses a modified branch of SANTA-SIM that exports the true fitness of each sample
in the FASTA header of the alignment output. This commit also makes SANTA-SIM
emit samples more often between generations to get the effect of multiple sample
timepoints within a single month. This change should be helpful for Nextstrain
pipelines that are used to working with real (noisier) data.