Add waning effect to cross-immunity calculations
Defines a longer period to calculate pairwise distances between samples and also
implements a linear waning effect on the cross-immunity cost per sample. As in
Bell et al. [1], this effect scales the cross-immunity by the number of years
between two samples divided by the waning period. After this waning period, the
cross-immunity cost goes to zero. The goal of this approach is to prevent
samples from the far past from dominating the cross-immunity metric when more
recent past samples are more likely to be biologically relevant.

[1] https://bedford.io/pdfs/papers/bell-dengue-antigenic-dynamics.pdf