Set L1 lambda such that true fitness is never less accurate than naive model
We would like an L1 lambda that helps us avoid model overfitting across training
windows. Since the mean distance to the future can always be minimized by
increasing coefficients to infinity, we need another metric to determine an
appropriate L1 constraint. Here, we choose the L1 value such that the true
fitness of a simulated population always performs better than the naive
model. The logic here is if we identify a model that approximates true fitness
for natural populations, the same relationship with the naive model is likely to
be true.