Fix a bug in how dates were annotated for simulated samples
Simulated samples are assigned floating point dates based on their generation in
the simulation plus a fixed offset of 2000 years. This number needs to be
converted to a standard datetime string of the format YYYY-MM-DD for downstream
steps of the pipeline. However, the `float_to_datestring` function that made
this conversion had a bug where samples with dates like 2021.28 that should have
been assigned a date string of 2021-04-12 were getting their date strings
rounded down to 2021-04-01. This caused strange behavior like samples not
getting frequencies estimated even though they were in the tree and other
inference problems.

This commit reverse engineers the logic of TreeTime's `numeric_date` utility [1]
to get a date string that could be converted back to the same floating point
value by this utility.

[1] https://github.com/neherlab/treetime/blob/2a1e819/treetime/utils.py#L142-L161