Make inclusion of cross-validation scores optional in model output
Adds a new flag to model fitting to opt-in for cross-validation scores to be
included in the model JSON output. By default, scores will not be included such
that the model JSON will be a minimal file with just the coefficients and
parameters used to fit the model (e.g., L1 lambda, etc.). This minimal file will
be the input for day-to-day forecasting in seasonal flu builds and
elsewhere. The complete file with cross-validation scores will be used for model
testing and inspection.