Decouple builds from datasets
Adds support for a new "datasets_sample" key in each build definition that
allows a given build to reuse an existing dataset that has already been built
under a different name. In the absence of a "datasets_sample" name in the build
config, the build's sample name will be used to lookup the dataset sample. This
change effectively decouples builds from datasets. One benefit of this
decoupling is that it allows us to test out new features on an existing dataset
with different build parameters and compare the results.