Add less background-dependent DMS predictor metrics
Adds two new distance maps based on DMS data: 1) a DMS "entropy" map which
weights distances between sequences based on the standardized entropy of the
site where the mutations occur. This implementation should produce a positive
coefficient such that mutations at sites with higher entropy are beneficial and
mutations at sites with low entropy are detrimental. 2) a DMS non-epitope map
which designates all sites with a standardized entropy less than zero as
non-epitope sites and counts the distance between sequences with a weight of 1
at these sites. We use both maps here relative to the ancestor from the last
season, to enable comparison of these metrics with the existing non-epitope and
DMS mutation effects metrics. The two new metrics should be much less dependent
on the background strain used for the DMS experiments than the DMS mutation
effects metric.