Fix a time boundary bug in partitioning of strains by timepoint
Fixes a floating point rounding bug that allowed strains with collection dates
equal to the current timepoint to be included in a timepoint's build even though
it would not get an estimate frequency or be included in the downstream
analysis. The fix here is to round the numeric dates passed to the partitioning
script to a monthly floating point resolution instead of the daily resolution of
the TreeTime numeric_date function.