Reduce run time for pairwise titer tree distances
Reduces the maximum number of past samples to consider for comparison to current
samples and also introduces a minimum frequency that a sample must be or have
been in the present or past to be included in a calculation. This latter change
helps us avoid calculations for samples that do not actually contribute to
forecasts.