Calculates tip frequencies with the appropriate minimum date
Previous to this commit, tip frequencies were calculated based on the date of
the earliest node in the given tree which was always the inferred MRCA of all
tips in the given build. What we want instead is to estimate frequencies just
for the years that we built the tree for. This commit explicitly takes this
number of years to build the tree into account when calculating the minimum date
for frequency estimation.