Rescale time for simulated populations
These changes increase the weighted Hamming distance between adjacent years to
more closely match the distance between natural populations. The primary goal
here is to make simulated populations and models we fit to them more comparable
with natural populations. With this rescaling, the penalty for
exposure-dependent fitness should not need to be as high as it was for the
densely sampled simulated population.