Require similar number of years per timepoint tree in builds
Improves model consistency by fitting models to similar amounts of data at each
timepoint (e.g., 6 years of data). Previously, we allowed input data to come
from timepoints with only 1 year of data. In simulations, we observed that this
configuration changed the density of tips compared between adjacent timepoints
in the earliest part of the analysis and led to less consistent model accuracies
and coefficients.

Since increasing the minimum number of years per build also reduces the number
of validation timepoints available for the same fixed analysis window, this
commit also increases the number of years simulated for sparse simulated
populations from 40 to 50 and uses 30 years of these data for training and
validation.