Add rule to forecast tips from all timepoints
Uses the best model for a given build to create a table of forecasts for all
tips at all timepoints. This table is used for model validation/testing. This
commit also modifies the forecasting script to make frequency inputs and node
data and frequencies outputs optional. The script will complain loudly with an
error if the user doesn't request any of the three possible inputs.