Test models on new data and export results to standard tables
This commit allows test results for models to be merged with validation results
into a single table. These results are distinguished by a new column
"error_type" that contains either "validation" or "test" depending on the
dataset. The primary benefit of merging these tables is for plotting the errors
per timepoint for each model across both validation and test sets.