Improved aesthetics for model diagram figure
Changes the color palette of the sequence-based clusters to a qualitative
palette from a quantitative one to avoid interpretation of the tip colors as a
continuous trait. Also reduces the alpha transparency and line width for circles
used to represent tips to address occlusion of data all mapping to the same part
of the plot.