Add epitope mutations by trunk status for simulated populations
To better understand why the epitope cross-immunity metric is not strongly
predictive of simulated flu evolution by itself, this commit includes the same
analysis I performed earlier for the natural populations where we count the
number of epitope and non-epitope mutations per branch based on the trunk or
side branch status of those branches. The resulting table gives us a sense of
how the relative importance of epitope mutations on the trunk vs. the side
branches (which we expect to be high) and also the absolute number of these
epitope mutations.