Address confusion about model selection metrics for simulations
Reformats the model selection table for simulation results to report
coefficients, distances to the future, approximations of the future's diversity,
and how often the models are better than the naive model. The naive model is now
explicitly included in this table. The combination of these reported metrics
allows the text to be updated to clarify how much better some models are than
others based on different targets (absolute distance to future, diversity of
future, etc.).

This commit also updated the model performance figures for simulated
populations, splitting the true fitness metric out to its own figure from the
other individual and best composite metrics and changing the reported metric
back to the absolute distance to the future with the naive model shown as a
reference for each timepoint.