Switch text and figures to absolute distances to the future
It is easier to communicate absolute distances to the future in amino acids
instead of values relative to the naive model. This commit includes text
describing how well models approximate future diversity, to address Richard's
comment about how we know whether any of the absolute distances to the future
are better than others.

This commit also includes an expanded discussion of the epitope cross-immunity
results for simulated and natural populations to address why these metrics are
less predictive than we expect (and for different reasons).

Finally, this commit includes a new supplemental table of test results for
simulated populations mirroring the validation results for the same
populations. This table is helpful for seeing how different models perform on
truly out-of-sample data.