Add references to trunk/side branch tables, remove DMS result figure
Tables with mutations in trunk vs. side branch are useful and should be included
as evidence of how the simulated populations match the natural
populations. There is no need for a separate figure showing DMS metric
coefficients and errors through time since we already have a table that
summarizes these.