Update outliers
Primarily excludes Shiraz strains that have collection dates in 2015 and 2016
but that are most closely related to samples from 2011-2012. This mismatch in
genetic distance and annotated collection date produces branches that span 3-4
years with an apparent lack of any amino acid mutations during that
period. Since this pattern doesn't match the evolutionary clock of H3N2, we
exclude these samples and a handful of similarly mismatched samples.