Annotate model JSONs with tip attributes
Collects redundant logic for annotating model JSONs into a Python script that
also supports annotating models by tip attribute data. Specifically, we annotate
the weighted distance to the future for each timepoint. This information is not
available for all timepoints in the validation/test scores, so we add it here.