Add analysis for historical contingency of epitope sites
Includes a description of and results for a new fitness metric, oracle
cross-immunity, based on knowledge from the future of which sites contain
beneficial mutations across the entire training/validation period. Results
include a supplemental figure comparing epitope cross-immunity and oracle
cross-immunity coefficients and distances through time and a new line in the
complete table of models fit to natural populations.

Based on the results of this new metric, this commit also substantially rewrites
the description of antigenic model performance for natural populations. The
general conclusion of this new section is that the epitope cross-immunity metric
can be extremely effective given that the "epitope" sites it uses are known in
advance.