Correct test results for simulated population with 18 timepoints
Previous results were based on an accidental inclusion of two timepoints in the
test data that did not actually include any new simulated samples. This issue
was caused by the fact that I only simulated 10,000 generations or 50 years of
data, but the assumption of 10 years of test timepoints would have required 51
years of data. This commit updates the test results based on the 18 valid
timepoints.