Add four-panel validation figure with clade errors
Adds a new Jupyter notebook with the simplest logic required to plot a single
four-panel validation figure for simulated populations. This figure includes two
additional panels showing absolute clade errors with a LOESS fit and the
correlation of all estimated and observed ranks for all strains across all
timepoints.

Also adds a generic script version of the above notebook that takes the
appropriate input files and outputs one of the validation figures for the
corresponding data. This script will allow these figures to be automatically
produced from the Snakemake rules for each population. This script and the above
notebook perform aggregation of a pandas Series that requires pandas version
0.25.0 or later, so the environment YAML file has also been updated here.

Finally, adds the updated validation figure for simulated populations.