Add comparison of vaccine strains with observed/estimated best
Includes a more detailed JSON of vaccine strains for H3N2 based on the seasonal
flu build JSON of the same name, adding earlier strains that were missing from
live seasonal flu builds and other metadata included which forecasting timepoint
each selection date corresponds to and the amino acid sequence of each vaccine
strain. These last two components allow the notebook in this commit to calculate
weighted distances to the future for vaccine strains even if those strains were
not included in the original natural population builds (due to filtering of
egg-passaged sequences).

The attached notebook plots the weighted distance to the future for each vaccine
strain compared to the observed closest strain to the future and our models'
estimates across the validation and test time periods.