Compare vaccine strains with observed and estimated best strains
Adds a figure comparing the vaccine strain selections per timepoint to the
observed closest strain to the future and the models' estimated closest
strain. Also, updates the corresponding Jupyter notebook to resize the image to
so the text annotations don't get cropped from the final image on the right side.

Closes #59.