Update results for best model in natural population test data
After updating to the model with the absolute lowest mean distance to the future
in the test data, this commit updates the corresponding text and figures to
reflect this model's performance. This commit also corrects a typo in the rank
results for the test data caused by a copy/paste of prose.