Fix a subtle bug in the DMS entropy distance map
Fixes an issue where the weighted distance between sequences from the DMS
entropy map was converted from a float to an integer because the default value
was defined as an integer 0 instead of 0.0. This is obviously an edge case bug
in augur distance that I need to fix, but in the short term, this change will
address the issue for this analysis.