Include FRA background and results
Restructures main text to introduce FRA assays early on as a more recent
alternative to HI assays, describe the limited availability of FRA data for
training/validation/testing along HI models, include FRA results in the final
results section, and reduce the description of FRAs as future directions in the
discussion. This commit also include an additional citation for the
glycosylation motifs in modern H3N2s (Chambers et al. 2015) and a citation for
the FRAs (Okuno et al. 1990).