Add FRA cross-immunity predictor to builds
This requires some duplication of Snakemake rules, for now, to download titers,
fit a titer model with FRA data instead of HI data, and handle all of the
downstream steps to calculate cross-immunty from FRA pairwise tree
distances. Part of the changes here include a modification to the pairwise
distance script to allow the user to specify which attribute to use from the
titer model JSON for pairwise distance calculations.