Merge edits after clearing CDC review
One major change here is the clarification that forecasts made by the model from
January 2020 to 2021 were made on January 17. This addresses a concern of Xiyan
Xu that A1b/131K as it is currently circulating would need to accumulate more
mutations to be successful. Relative to the perspective of January 17, these
changes may have already happened by May 16. I added a brief description of how
A1b/131K and 3c3.A have changed in frequency in the four months since those
initial predictions.