Add bootstrap analysis to text and tables
Updates Tables 1 and 2 and Supplemental Table S3 and their captions to include
indicators for models that were significantly closer to the future than the
corresponding populations' naive model. Adds a description of the bootstrap
hypothesis tests to the methods and adds figures displaying actual p-values and
corresponding bootstraps of different models to the supplementary material.