Calculate lower bound for EMD between validation and test timepoints
When calculating the distance to the future for the naive model, also calculated
the lowest bound for the EMD between each timepoint and the future. We
calculate the lower bound using a greedy optimization strategy that maps each
strain in the future to its closest strain by Hamming distance in the current
timepoint and assigns each future strain's frequency to its corresponding
current strain's estimated frequencies. Then we calculate the EMD between these
new estimated frequencies for the current timepoint strains and the observed
future frequencies and existing distance matrix between strains. The resulting
EMD should be the smallest distance to the future any model could hope to
calculate since it maps all information about future strains to the current
timepoint.