Enable running natural builds with frozen list of strains
Strain select can change stochastically and with the addition of newest data in
our fauna database. To prevent strains from the analysis from constantly
changing, allow the user to define a path to a list of frozen strain names to
use in a given analysis. The same approach is not currently possible for the
simulated sequences due to randomness is subsampling of these sequences. That
randomness is addressed by a more recent version of augur than we are using
here.