Add an analysis of log2 distances to the future
Attempts to calculate log2 distances to the future for strains included in the
vaccine comparison figure. After initial prototyping, I determined that we did
not have all of the data we needed to perform this analysis without completely
re-running the entire workflow after including four missing vaccine strains. It
was also not clear whether the relationship between Hamming distance and log2
titer distance was something we could investigate without more even titer
coverage for each strain. This commit includes this work, in case we want to
revisit it for a future study.