Update significance annotations based on new bootstrap tests
Bootstrap tests now use the difference between each biologically-informed
model's distance and the naive model's distance at each timepoint as the
empirical distribution for bootstrap hypothesis testing. This change attempts to
account for timepoint-to-timepoint variation where the absolute distances
between populations may vary widely but the actual difference between models
does not.

This commit includes new prototypes of visual displays for the bootstrap testing
that more accurately represent the origin of the p-values we calculate. Final
versions of these plots will appear in a future commit.

Also note that the bootstrap p-values here will vary slightly from run to run
due to random sampling effects, but the conclusions we make from these values
should be generally robust to that variation (i.e., values are unlikely to be
"almost significant" in one run and "significant" in another).