Simplify bootstrap analysis
After switching to bootstraps of differences between biologically-informed
models and the corresponding naive models, there were only one-sample
distributions to test against a null hypothesis that these values were not
zero. My initial attempts to use bootstrap hypothesis tests with a null
distribution shifted to a mean of zero appear to be less justified after reading
more of the literature. These tests are also more complicated to interpret and
explain.

Here, I've replaced these tests with a simple p-value estimate based on the
proportion of bootstrapped samples of model differences that were less than zero
out of 10,000 bootstraps. This test has its weaknesses, too, in that it likely
overestimates the significance of differences by not accounting for variance of
the null model and also by effectively using bootstrap confidence intervals
calculated from small sample sizes (n=8 for the test period for natural
populations) which are known to be biased toward smaller values. These new
results represent a trade-off in complexity and accuracy.

Original analyses by bootstrap hypothesis testing are included here for
posterity.