Clarify bootstrap p value results
Set the y-axis range of the bootstrap distribution figures to the same values
for both validation and test data to more clearly represent the higher variance
of the test data points for natural populations. Also update the methods text
for bootstrap tests to explicitly state the meaning of the p value.