Prevent Snakemake from downloading natural sequences by default
Natural sequences are only available through GISAID or an individual user's or
lab's curated database of sequences. This commit only allows data to be
downloaded if the user has provided proper credentials to access an internal
database and otherwise alerts the user to download data from GISAID.

This commit also updates the documentation for downloading GISAID data such that
the user's sequence and metadata have the same names needed to start the
workflow for natural sequences.