Clarify and expand comparison of models to vaccine strains
Adds best estimated strains by naive model to vaccine comparison figure, adds a
new supplemental figure that shows the distance to the future of observed and
estimated best strains relative to the corresponding vaccine strains, and
quantifies improvement of models over the vaccine strains in the results text.

This commit also changes the vaccine strain name annotation in the vaccine
comparison figure to use an abbreviation with a larger font and improve overall
readability. The abbreviations are now explained in the figure caption.