Summarize antigenic distances in vaccine comparison
Adds analyses to the vaccine comparison notebook that include a comparison of
relative distance to the future for closest strains vs. vaccine strains and
the antigenic distance between those corresponding strains. Also, improves the
aesthetics and readability of the vaccine comparison figure and adds the
supplemental figure showing relative distance to the future for closest strains
vs. vaccine strains.