Replace "best strain" with "closest strain" and explain
Replaces the "best" value judgement with the quantitative "closest" for each
closest strain to the future, explains how we calculate these closest strains to
the future the first time they are mentioned, and references the equations
used. Also expands the interpretation of these closest strains in the vaccine
comparison section to clarify that they represent centroids of the observed and
estimated future populations.