Add initial code to calculate distance from consensus metric
Adds a predictor entry in the config for simulated populations, a rule to the
Snakemake workflow, and a script to calculate the consensus sequence from
strains with non-zero frequencies at a given timepoint and the corresponding
distance from the consensus for all of those same strains. These values should
get merged into the final tip attributes table such that the
"distance_from_consensus" attribute can be used to fit a model for simulated and
natural populations.