Update instructions to run the workflow
Modernizes instructions to install Snakemake and run the workflow.
Instead of manually installing the workflow's Conda environment and
using the Snakemake in that environment to run the workflow, install a
modern Snakemake and let Snakemake create the workflow environment on
the fly. If the user has installed a modern Miniconda, the Conda solver
will be using libmamba which is substantially faster than the old Conda
solver that gets installed in the workflow's environment. Snakemake
still tries to use mamba as the default Conda backend, so this commit
updates all Snakemake commands to explicitly set the backend to "conda".